프로파일 나누기 및 서비스 코드 완성

+++ post/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
 
-from post.models import Post
+from post.models import Post, Comment
 
 
 # 게시글 목록 페이지
@@ -13,17 +13,19 @@ def index(request):
 # 게시글 상세 페이지
 def detail(request, post_id):
     post = get_object_or_404(Post, post_id=post_id)
-    context = {'post': post}
+    comment_list = Comment.objects.filter(post=post).order_by('-comment_id')
+    context = {'post': post, 'comments': comment_list}
     return render(request, 'post/detail.html', context)
 
 
 # 게시글 등록 페이지
 def create_post(request):
+    if not request.user.is_authenticated:
+        return redirect('login:index')
     if request.method == "POST":
         title = request.POST.get("title")
-        author = request.POST.get("author")
         content = request.POST.get("content")
-        post = Post(title=title, author=author, content=content)
+        post = Post(title=title, user=request.user, content=content)
         post.save()
         return redirect('post:detail', post_id=post.post_id)
     return render(request, 'post/create.html')
@@ -31,19 +33,40 @@ def create_post(request):
 
 # 게시글 삭제
 def delete_post(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('login:index')
     if request.method == "GET":
         post = get_object_or_404(Post, post_id= post_id)
-        post.delete()
-        return redirect('post:index')
+        if post.user.username != request.user.username:
+            return redirect('post:detail', post_id=post.post_id)
+        else:
+            post.delete()
+            return redirect('post:index')
 
 # 게시글 수정
 def update_post(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('login:index')
     post = get_object_or_404(Post, post_id=post_id)
+    if post.user.username != request.user.username:
+        return redirect('post:detail', post_id=post.post_id)
     if request.method == "POST":
         post.title = request.POST.get("title")
-        post.author = request.POST.get("author")
         post.content = request.POST.get("content")
         post.save()
         return redirect("post:detail", post_id=post_id)
     context = {'post': post}
     return render(request, 'post/update.html', context)
+
+# 댓글 추가
+def create_comment(request, post_id):
+    if not request.user.is_authenticated:
+        return redirect('login:index')
+    post = get_object_or_404(Post,post_id=post_id)
+    if request.method == 'POST':
+        comment_content = request.POST.get('comment')
+        comment_user = request.user
+        comment = Comment(comment_content = comment_content, comment_user = comment_user, post=post)
+        comment.save()
+        return redirect('post:detail', post_id = post_id)
+
